Handle repository-scoped attributes when transferring records between repositories
  * If we transfer all records attached to an assessment, we transfer
    the assessment as well.

  * If we transfer a subset of records, we clone the assessment and
    create an equivalent one in the target repository.

If the assessment contained custom repository-specific attributes,
this would mean that the assessment record in the target repository
might be missing data.  To avoid that, we look for an attribute of the
same type and with the same label in the target repository and use
that if we find it.