Move notes into a separate table
Now that ratings can be set to "Not set", we could end up in a
situation where there's an `assessment_attribute` row that links to a
definition but has a NULL for both `value` and `note`.  This would
count as a "use" of the definition, even though it isn't really.

We were also in a weird situation where "not set" could be represented
in two different ways: either by having a NULL in the value column, or
by having no row at all in assessment_attribute.

Normalization to the rescue!  Move the note column into its own table,
and only put a row in there if there's actually a note.  Likewise,
only put a row in `assessment_attribute` if there's a value set.