Ensure that realtime-indexing always sees up-to-date records
While developing a plugin, we noticed a bug with the realtime indexing
whereby the records sent to Solr were missing their notes.  This is
hard to notice because the periodic indexer will fix up the index
around 30 seconds later, but if you happened to search on notes
content in that 30 second window you wouldn't find what you were
looking for.

The bug stems from the fact that updated records get emitted for
realtime indexing after the database transaction commits, but get
prepared for being emitted *before* the commit happens.  As all
database updates might not have been applied at this point, partial
reads are possible.

The fix is to pull the record from the database at the point it gets
sent for realtime indexing--after the commit has finished.  The fact
that this wasn't happening already was likely just an oversight from
me, as the original realtime indexing code didn't pull records from
the database at all, so this was never an issue in the original code.

These changes are unfortunately difficult to test with standard
backend unit tests, as the DB transaction lifecycle is modified in the
unit test environment.  I've written an integration test to exercise
the update feed instead, and this works as expected.