When bumping, only update the lock_version if the system_mtime is at least a second old.
Still bump the system_mtime though. This allows the indexer to still do its thing,
while the lock_version bump has already happened so other editors will still get a
version conflict.

The motivation is to reduce the number of meaningless versions being created
when a bulk update function does a bunch of updates at once (like MoG).

Normally this would go unremarked upon, but if you happen to be using the
as_history plugin (which bases its versioning on lock_version), you will see
potentially a lot of versions all from the same user at the same time with
no diffs - which is boring and confusing at the same time. Terrible.

This is UNTESTED and can probably be written in a better way, and the method
should probably be renamed.

At the moment, mostly interested in the viability of the concept, and how
people 'feel' about it.