Piv#172937068 Handle possessive apostrophes
If we have a document with a title like:

  Coyle's Postwar

we want the following searches to find it:

  * coyle's postwar
  * coyle postwar
  * coyles postwar

as well as phrased versions of the same queries.

I couldn't find a combination of existing filters that would do this,
so here's a custom token filter that generates the right variations of
these terms.